Tweaked scale, origins, and fixed animation bug at extreme Y values.

+++ main/modes/games/bigbug/entityManager_bigbug.h
@@ -50,7 +50,7 @@ void bb_drawEntities(bb_entityManager_t* entityManager, rectangle_t* camera);
 bb_entity_t* bb_findInactiveEntity(bb_entityManager_t* entityManager);
 
 void bb_viewFollowEntity(bb_entity_t* entity);
-bb_entity_t* bb_createEntity(bb_entityManager_t* entityManager, uint8_t type, uint8_t spriteIndex, uint16_t x, uint16_t y);
+bb_entity_t* bb_createEntity(bb_entityManager_t* entityManager, uint8_t type, uint8_t spriteIndex, uint32_t x, uint32_t y);
 
 void bb_freeEntityManager(bb_entityManager_t* entityManager);
